Updates/roles (#2)
* Updated roles to latest


* Updated info about repo

+++ roles/ansible-consul/Vagrant/bootstrap.sh
@@ -13,11 +13,21 @@ if [ -f /etc/debian_version ]; then
     sudo apt-get update && \
     sudo apt-get -y install python-simplejson
   fi
+  if [[ $codename == "yakkety" ]]; then
+    sudo apt-get update && \
+    sudo apt-get -y install python-simplejson
+  fi
+  if [[ $codename == "zesty" ]]; then
+    sudo apt-get update && \
+    sudo apt-get -y install python-simplejson
+  fi
 fi
 if [ -f /etc/redhat-release ]; then
-  codename="$(gawk -F= '/^NAME/{print $2}' /etc/os-release)"
-  if [[ $codename == "Fedora" ]]; then
-    sudo dnf -y install python-devel python-dnf && \
-    sudo dnf -y group install "C Development Tools and Libraries"
+  if [ -f /etc/os-release ]; then
+    codename="$(gawk -F= '/^NAME/{print $2}' /etc/os-release)"
+    if [[ $codename == "Fedora" ]]; then
+      sudo dnf -y install python-devel python-dnf && \
+      sudo dnf -y group install "C Development Tools and Libraries"
+    fi
   fi
 fi
